backend/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
sentence-transformers==3.0.1
rank-bm25==0.2.2
pinecone-client==3.2.2
openai>=1.51.0
pandas==2.2.2
pydantic>=2.9.0
